Move annotations functionality to its own file
